+++ ui/src/MetadataPage.vue
@@ -20,6 +20,8 @@ import {
 import {
   AgGridTheme,
   makeRowNumberColDef,
+  PlAgColumnHeader,
+  type PlAgHeaderComponentParams,
   PlAgOverlayNoRows,
   PlBlockPage,
   PlBtnGhost,
@@ -139,7 +141,9 @@ const columnDefs = computed<ColDef[]>(() => [
     headerName: app.model.args.sampleLabelColumnLabel,
     initialWidth: 200,
     flex: 1,
-    suppressHeaderMenuButton: true
+    suppressHeaderMenuButton: true,
+    headerComponent: PlAgColumnHeader,
+    headerComponentParams: { type: 'Text' } satisfies PlAgHeaderComponentParams
   },
   {
     colId: 'datasets',
@@ -154,7 +158,9 @@ const columnDefs = computed<ColDef[]>(() => [
     }),
     // flex: 1,
     suppressHeaderMenuButton: true,
-    width: 200
+    width: 200,
+    headerComponent: PlAgColumnHeader,
+    headerComponentParams: { type: 'Text' } satisfies PlAgHeaderComponentParams
   },
   ...app.model.args.metadata.map((mCol): ColDef => {
     const common: ColDef = {
@@ -166,12 +172,18 @@ const columnDefs = computed<ColDef[]>(() => [
     };
     switch (mCol.valueType) {
       case 'String':
-        return common;
+        return {
+          ...common,
+          headerComponent: PlAgColumnHeader,
+          headerComponentParams: { type: 'Text' } satisfies PlAgHeaderComponentParams
+        };
       case 'Double':
         return {
           ...common,
           cellDataType: 'number',
-          cellEditor: 'agNumberCellEditor'
+          cellEditor: 'agNumberCellEditor',
+          headerComponent: PlAgColumnHeader,
+          headerComponentParams: { type: 'Number' } satisfies PlAgHeaderComponentParams
         };
       case 'Long':
         return {
@@ -181,7 +193,9 @@ const columnDefs = computed<ColDef[]>(() => [
           cellEditorParams: {
             precision: 0,
             showStepperButtons: true
-          }
+          },
+          headerComponent: PlAgColumnHeader,
+          headerComponentParams: { type: 'Number' } satisfies PlAgHeaderComponentParams
         };
     }
   }),
@@ -221,7 +235,11 @@ const rowData = computed<MetadataRow[]>(() => {
 const gridOptions: GridOptions<MetadataRow> = {
   getRowId: (row) => row.data.id,
 
-  rowSelection: 'multiple',
+  rowSelection: {
+    mode: 'multiRow',
+    checkboxes: false,
+    headerCheckbox: false
+  },
 
   autoSizeStrategy: {
     type: 'fitGridWidth'
@@ -291,10 +309,11 @@ const gridOptions: GridOptions<MetadataRow> = {
     if (getSelectedSamples(params.node).length === 0) return [];
     return [
       {
-        name: `Delete ${targetSamples.length > 1
-          ? `${targetSamples.length} samples`
-          : app.model.args.sampleLabels[targetSamples[0]]
-          }`,
+        name: `Delete ${
+          targetSamples.length > 1
+            ? `${targetSamples.length} samples`
+            : app.model.args.sampleLabels[targetSamples[0]]
+        }`,
         action: (params) => {
           const samplesToDelete = getSelectedSamples(params.node);
           deleteSamples(targetSamples);
@@ -312,33 +331,50 @@ const gridOptions: GridOptions<MetadataRow> = {
 <template>
   <PlBlockPage>
     <template #title>
-      <PlEditableTitle max-width="600px" placeholder="Samples & Data" :max-length="40"
-        v-model="app.model.args.blockTitle" />
+      <PlEditableTitle
+        v-model="app.model.args.blockTitle"
+        :max-length="40"
+        max-width="600px"
+        placeholder="Samples & Data"
+      />
     </template>
     <template #append>
       <PlBtnGhost @click.stop="() => (app.showImportDataset = true)" icon="dna-import">
         Import sequencing data
       </PlBtnGhost>
       &nbsp;
-      <PlBtnGhost @click.stop="importMetadata" icon="table-import">
-        Import metadata
-      </PlBtnGhost>
+      <PlBtnGhost icon="table-import" @click.stop="importMetadata"> Import metadata </PlBtnGhost>
     </template>
     <div :style="{ flex: 1 }">
-      <AgGridVue :theme="AgGridTheme" :style="{ height: '100%' }" @grid-ready="onGridReady" :rowData="rowData"
-        :columnDefs="columnDefs" :grid-options="gridOptions" :noRowsOverlayComponent="PlAgOverlayNoRows" />
+      <AgGridVue
+        :theme="AgGridTheme"
+        :style="{ height: '100%' }"
+        :rowData="rowData"
+        :columnDefs="columnDefs"
+        :grid-options="gridOptions"
+        :noRowsOverlayComponent="PlAgOverlayNoRows"
+        @grid-ready="onGridReady"
+      />
     </div>
   </PlBlockPage>
 
   <ImportDatasetDialog v-if="app.showImportDataset" />
 
-  <ImportMetadataModal v-if="data.importCandidate !== undefined" :import-candidate="data.importCandidate"
-    @on-close="data.importCandidate = undefined" />
+  <ImportMetadataModal
+    v-if="data.importCandidate !== undefined"
+    :import-candidate="data.importCandidate"
+    @on-close="data.importCandidate = undefined"
+  />
 
-  <PlDialogModal :model-value="data.errorMessage !== undefined" closable @update:model-value="(v) => {
-    if (!v) data.errorMessage = undefined;
-  }
-    ">
+  <PlDialogModal
+    :model-value="data.errorMessage !== undefined"
+    closable
+    @update:model-value="
+      (v) => {
+        if (!v) data.errorMessage = undefined;
+      }
+    "
+  >
     <div>{{ data.errorMessage?.title }}</div>
     <pre v-if="data.errorMessage?.message">{{ data.errorMessage?.message }}</pre>
     <PlBtnPrimary>Ok</PlBtnPrimary>