+++ ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-open/milaboratories.samples-and-data.ui
 
+## 1.16.2
+
+### Patch Changes
+
+- 31a78a6: Added column type marker in header
+
+## 1.16.1
+
+### Patch Changes
+
+- d5a1df7: package version update
+
 ## 1.16.0
 
 ### Minor Changes